requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
PyYAML
